size_t

+++ google/cloud/pubsub/internal/publisher_tracing_connection_test.cc
@@ -88,7 +88,7 @@ TEST(PublisherTracingConnectionTest, PublishSpanOnSuccess) {
                     OTelAttribute<std::string>("messaging.pubsub.ordering_key",
                                                "ordering-key-0"),
                     OTelAttribute<int>("gcloud.status_code", 0),
-                    OTelAttribute<std::uint32_t>(
+                    OTelAttribute<std::size_t>(
                         "messaging.message.total_size_bytes", 45),
                     OTelAttribute<std::string>("messaging.message_id",
                                                "test-id-0")))));
